Ugrh!
You know, on the night of the 14th, I began to get very annoyed and enrage what I was doing, BUT, at 7 a.m. I got a brilliant idea how the code works, and I could explain in my mind how I really need to write this code. Fuhhh I'm starting to despise ...